requirements.txt
pytesseract
Pillow
playwright